+++ builtin/providers/azurerm/resource_arm_loadbalancer_nat_pool.go
@@ -51,8 +51,10 @@ func resourceArmLoadBalancerNatPool() *schema.Resource {
 			},
 
 			"protocol": {
-				Type:     schema.TypeString,
-				Required: true,
+				Type:             schema.TypeString,
+				Required:         true,
+				StateFunc:        ignoreCaseStateFunc,
+				DiffSuppressFunc: ignoreCaseDiffSuppressFunc,
 			},
 
 			"frontend_port_start": {